Update FEMTO_ST.md
I recently found a mistake in my upload of the dataset. All the temperature data was missing from the training data. I've uploaded this, but your linked fork of my repository hasn't updated for a while. To guarantee the correct dataset, I think it's a good idea to change the link.